file deleting is done

+++ Python Advanced/13 - File Handling/04. file_delete.py
@@ -1,11 +1,22 @@
 # First task from the lecture
+# First option
 
 import os
 
 path = os.path.join("resources", "numbers.txt")
-os.remove(path)
 
-
+try:
+    os.remove(path)
+except FileNotFoundError:
+    print("The file is already deleted")
+    
+# Second option
+        
+if os.path.exists(path):
+    os.remove(path)
+else:
+    print("The file is already deleted")
+    
 # Second task from me
